+++ mycodo/config.py
@@ -821,10 +821,6 @@ FUNCTION_ACTIONS = [
 
 # Calibration
 CALIBRATION_DEVICES = [
-    ('setup_atlas_ec', 'Atlas Scientific Electrical Conductivity Sensor'),
-    ('setup_atlas_ezo_pump', 'Atlas Scientific EZO Pump'),
-    ('setup_atlas_ph', 'Atlas Scientific pH Sensor'),
-    ('setup_atlas_rgb', 'Atlas Scientific RGB Sensor'),
     ('setup_ds_resolution', 'DS-Type Temperature Sensors (e.g. DS18B20)')
 ]