+++ mycodo/inputs/atlas_ec.py
@@ -74,12 +74,70 @@ INPUT_INFORMATION = {
         }
     ],
 
-    'custom_actions_message':
-        'The I2C address can be changed. Enter a new address in the 0xYY format '
-        '(e.g. 0x22, 0x50), then press Set I2C Address. Remember to deactivate and '
-        'change the I2C address option after setting the new address.',
-
     'custom_actions': [
+        {
+            'type': 'message',
+            'default_value': """Calibration: a one- or two-point calibration can be performed. It's a good idea to clear the calibration before calibrating. Always perform a dry calibration with the probe in the air (not in any fluid). Then perform either a one- or two-point calibration with calibrated solutions. If performing a one-point calibration, use the Single Point Calibration field and button. If performing a two-point calibration, use the Low and High Point Calibration fields and buttons. Allow a minute or two after submerging your probe in a calibration solution for the measurements to equilibrate before calibrating to that solution. The EZO EC circuit default temperature compensation is set to 25 °C. If the temperature of the calibration solution is +/- 2 °C from 25 °C, consider setting the temperature compensation first. Note that at no point should you change the temperature compensation value during calibration. Therefore, if you have previously enabled temperature compensation, allow at least one measurement to occur (to set the compensation value), then disable the temperature compensation measurement while you calibrate."""
+        },
+        {
+            'id': 'clear_calibrate',
+            'type': 'button',
+            'name': 'Clear Calibration'
+        },
+        {
+            'id': 'dry_calibrate',
+            'type': 'button',
+            'name': 'Calibrate Dry'
+        },
+        {
+            'type': 'new_line'
+        },
+        {
+            'id': 'single_point_ec',
+            'type': 'integer',
+            'default_value': 84,
+            'name': 'Single Point EC (µS)',
+            'phrase': 'The EC (µS) of the single point calibration solution'
+        },
+        {
+            'id': 'single_calibrate',
+            'type': 'button',
+            'name': 'Calibrate Single Point'
+        },
+        {
+            'type': 'new_line'
+        },
+        {
+            'id': 'low_point_ec',
+            'type': 'integer',
+            'default_value': 12880,
+            'name': 'Low Point EC (µS)',
+            'phrase': 'The EC (µS) of the low point calibration solution'
+        },
+        {
+            'id': 'low_calibrate',
+            'type': 'button',
+            'name': 'Calibrate Low Point'
+        },
+        {
+            'type': 'new_line'
+        },
+        {
+            'id': 'high_point_ec',
+            'type': 'integer',
+            'default_value': 80000,
+            'name': 'High Point EC (µS)',
+            'phrase': 'The EC (µS) of the high point calibration solution'
+        },
+        {
+            'id': 'high_calibrate',
+            'type': 'button',
+            'name': 'Calibrate High Point'
+        },
+        {
+            'type': 'message',
+            'default_value': """The I2C address can be changed. Enter a new address in the 0xYY format (e.g. 0x22, 0x50), then press Set I2C Address. Remember to deactivate and change the I2C address option after setting the new address."""
+        },
         {
             'id': 'new_i2c_address',
             'type': 'text',
@@ -194,6 +252,50 @@ class InputModule(AbstractInput):
 
         return self.return_dict
 
+    def calibrate(self, level, ec):
+        try:
+            if level == "clear":
+                write_cmd = "Cal,clear"
+            elif level == "dry":
+                write_cmd = "Cal,dry"
+            elif level == "single":
+                write_cmd = "Cal,{}".format(level, ec)
+            else:
+                write_cmd = "Cal,{},{}".format(level, ec)
+            self.logger.debug("Calibration command: {}".format(write_cmd))
+            ret_val = self.atlas_device.write(write_cmd)
+            self.logger.info("Command returned: {}".format(ret_val))
+            # Verify calibration saved
+            write_cmd = "Cal,?"
+            self.logger.info("Device Calibrated?: {}".format(
+                self.atlas_device.write(write_cmd)))
+        except:
+            self.logger.exception("Exception calibrating")
+
+    def clear_calibrate(self, args_dict):
+        self.calibrate('clear', None)
+
+    def dry_calibrate(self, args_dict):
+        self.calibrate('dry', None)
+
+    def single_calibrate(self, args_dict):
+        if 'single_point_ec' not in args_dict:
+            self.logger.error("Cannot calibrate without calibration solution EC")
+            return
+        self.calibrate('single', args_dict['single_point_ec'])
+
+    def low_calibrate(self, args_dict):
+        if 'low_point_ec' not in args_dict:
+            self.logger.error("Cannot calibrate without calibration solution EC")
+            return
+        self.calibrate('low', args_dict['low_point_ec'])
+
+    def high_calibrate(self, args_dict):
+        if 'high_point_ec' not in args_dict:
+            self.logger.error("Cannot calibrate without calibration solution EC")
+            return
+        self.calibrate('high', args_dict['high_point_ec'])
+
     def set_i2c_address(self, args_dict):
         if 'new_i2c_address' not in args_dict:
             self.logger.error("Cannot set new I2C address without an I2C address")
@@ -202,6 +304,8 @@ class InputModule(AbstractInput):
             i2c_address = int(str(args_dict['new_i2c_address']), 16)
             write_cmd = "I2C,{}".format(i2c_address)
             self.logger.debug("I2C Change command: {}".format(write_cmd))
-            self.atlas_device.write(write_cmd)
+            ret_val = self.atlas_device.write(write_cmd)
+            self.logger.info("Command returned: {}".format(ret_val))
+            self.atlas_device = None
         except:
             self.logger.exception("Exception changing I2C address")